Crate wifilocation

A crate to return your GPS location using WiFi hotspots.

Usage

This crate is on crates.io and can be used by adding wifilocation to the dependencies in your project's Cargo.toml.

[dependencies]
wifilocation = "0.3.*"

and this to your crate root:

extern crate wifilocation;

Structs

GpsLocation
GPS struct to return longitude and latitude coordinates; and accuracy of information.

Location

Enums

Error

Functions

get_location
Return GPS location using a Vec of wifiscanner::Wifi. Uses Google's GPS location service

get_towers
A wrapper around wifiscanner scan function to return a Vec of wifiscanner::Wifi. The function was an artifact of the original node module and will probably not make